Do some refactoring for MapContainer
