+++ client/src/visualization/marker.tsx
@@ -4,6 +4,8 @@ import { Marker, Popup } from 'react-map-gl';
 
 import LocationPin from './locationPin';
 import { Location } from '../common/typeUtil';
+import { BeaconGeoLocation } from '../location/mqttDeserialize';
+import partition = require('lodash/partition');
 
 const StaticMarker = styled.div`
   svg {
@@ -80,4 +82,27 @@ export const LocationPinMarker = ({ type, coords, onClick }: PinProps) => {
     case 'none':
       return null;
   }
+};
+
+/**
+ * Why can there be multiple markers for the user? Because we cannot get unique
+ * Id for the device thanks to bluetooth security limits. Instead we can utilize
+ * the non-unique bluetooth name.
+ */
+export const divideMarkers = (
+  beacons: BeaconGeoLocation[],
+  bluetoothName: string | null,
+  lastKnownPosition: BeaconGeoLocation | null
+) => {
+  const [userMarkers, nonUserMarkers] = partition(
+    beacons,
+    beacon => beacon.beaconId === bluetoothName
+  );
+
+  const allUserMarkers =
+    lastKnownPosition && userMarkers.length === 0
+      ? [lastKnownPosition]
+      : userMarkers;
+
+  return { isOnline: userMarkers.length !== 0, allUserMarkers, nonUserMarkers };
 };
